+++ tests/spec_decode/e2e/test_correctness.py
@@ -595,137 +595,3 @@ def run_greedy_equality_correctness_test(baseline_llm_generator,
         print(f'{i=} {baseline_token_ids=}')
         print(f'{i=}     {spec_token_ids=}')
         assert baseline_token_ids == spec_token_ids
-
-@pytest.mark.parametrize(
-    "common_llm_kwargs",
-    [{
-        "model": "JackFram/llama-68m",
-
-        # Skip cuda graph recording for fast test.
-        "enforce_eager": True,
-
-        # Required for spec decode.
-        "use_v2_block_manager": True
-    }])
-@pytest.mark.parametrize("per_test_common_llm_kwargs", [{}])
-@pytest.mark.parametrize("baseline_llm_kwargs", [{}])
-@pytest.mark.parametrize(
-    "test_llm_kwargs",
-    [
-        {
-            "speculative_model": "JackFram/llama-160m",
-            "num_speculative_tokens": k,
-        }
-        # Try a range of common k, as well as large speculation.
-        for k in [2, 5]
-    ] + [
-        {
-            "speculative_model": "JackFram/llama-160m",
-            "num_speculative_tokens": 3,
-
-            # Artificially limit the draft model max model len; this forces vLLM
-            # to skip speculation once the sequences grow beyond 32-k tokens.
-            "speculative_max_model_len": 32,
-        }
-    ])
-@pytest.mark.parametrize("batch_size", [2, 8])
-@pytest.mark.parametrize(
-    "output_len",
-    [
-        # Use smaller output len for fast test.
-        32,
-    ])
-@pytest.mark.parametrize("seed", [1])
-def test_logprobs(baseline_llm_generator, test_llm_generator, batch_size: int,
-                output_len: int):
-    """Verify that speculative decoding produces exact equality to without spec
-    decode with many different values of k.
-    """
-    run_greedy_logprobs_correctness_test(baseline_llm_generator,
-                                         test_llm_generator,
-                                         batch_size,
-                                         max_output_len=output_len,
-                                         force_output_len=True)
-
-def run_greedy_logprobs_correctness_test(baseline_llm_generator,
-                                         test_llm_generator,
-                                         batch_size,
-                                         max_output_len,
-                                         force_output_len: bool,
-                                         logprob_rank: int = 1):
-    """Helper method that compares the outputs of both the baseline LLM and
-    the test LLM. It asserts greedy equality, e.g. that the outputs are exactly
-    the same when temperature is zero.
-    """
-    temperature = 0.0
-
-    prompts = [
-        "Hello, my name is",
-        "The president of the United States is",
-        "The capital of France is",
-        "The future of AI is",
-        "San Francisco is know for its",
-        "Facebook was created in 2004 by",
-        "Curious George is a",
-        "Python 3.11 brings improvements to its",
-    ]
-
-    prompts = [prompt for prompt, _ in zip(cycle(prompts), range(batch_size))]
-
-    # If the test requires that we generated max_output_len tokens, then set the
-    # sampling params to ignore eos token.
-    ignore_eos = force_output_len
-
-    sampling_params = SamplingParams(
-        max_tokens=max_output_len,
-        ignore_eos=ignore_eos,
-        temperature=temperature,
-        logprobs=logprob_rank,
-    )
-
-    #spec_batch_tokens, spec_batch_token_ids = get_logprobs_from_llm_generator(
-    #    test_llm_generator, prompts, sampling_params)
-    
-    spec_batch_logprobs = get_logprobs_from_llm_generator(test_llm_generator, prompts, sampling_params)
-    baseline_batch_logprobs = get_logprobs_from_llm_generator(baseline_llm_generator, prompts, sampling_params)
-
-    assert len(baseline_batch_logprobs) == len(prompts)
-    assert len(spec_batch_logprobs) == len(prompts)
-
-    for i, (baseline_logprobs, spec_logprobs) in enumerate(
-                zip(baseline_batch_logprobs, spec_batch_logprobs)):
-        assert len(spec_logprobs) == len(baseline_logprobs)
-        for pos, (spec_pos_logprobs, baseline_pos_logprobs) in enumerate(
-            zip(spec_logprobs, baseline_logprobs)):
-            
-            spec_rank_to_token_id = {
-                value.rank: key for key, value in spec_pos_logprobs.items()}
-            spec_rank_to_logprob = {
-                value.rank: value.logprob
-                for key, value in spec_pos_logprobs.items()}
-        
-            baseline_rank_to_token_id = {
-                value.rank: key for key, value in baseline_pos_logprobs.items()}
-            baseline_rank_to_logprob = {
-                value.rank: value.logprob
-                for key, value in baseline_pos_logprobs.items()}
-        
-            assert set(spec_rank_to_token_id.keys()) == set(baseline_rank_to_token_id.keys())
-
-            ranks = sorted(set(spec_rank_to_token_id.keys()))
-
-            for rank in ranks:
-                assert spec_rank_to_token_id[rank] == baseline_rank_to_token_id[rank]
-                #assert spec_rank_to_logprob[rank] == baseline_rank_to_logprob[rank]
-                if pos == 0 or True:
-                    import math
-                    #assert spec_rank_to_logprob[rank] == baseline_rank_to_logprob[rank]
-                    assert math.isclose(
-                        a=spec_rank_to_logprob[rank],
-                        b=baseline_rank_to_logprob[rank],
-                        #rel_tol=1e-3,
-                        abs_tol=1e-1,
-                    )
-                else:
-                    assert spec_rank_to_logprob[rank] == 0.1
-            #breakpoint()
